# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Log Service API",
    "version": "1.0.0",
    "description": "API documentation for the log ingestion, query, alerting and notification service."
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development server"
    }
  ],
...
  "components": {
...
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "connect.sid",
        "description": "Express session cookie established via POST /auth/login"
      }
    },
    "schemas": {
      "LogEntry": {
        "type": "object",
        "required": [
          "timestamp",
          "level",
          "service",
          "message"
        ],
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "description": "Log event timestamp. Rejected if more than 5 minutes in the future."
          },
          "level": {
            "type": "string",
            "enum": [
              "debug",
              "info",
              "warn",
              "error"
            ]
          },
          "service": {
            "type": "string",
            "description": "Name of the emitting service"
          },
          "message": {
            "type": "string",
            "description": "Log message"
          },
          "attributes": {
            "type": "object",
            "description": "Flat key/value attributes (string, number or boolean values)",
            "additionalProperties": {
              "type": "string"
            }
          }
        }
      },
      "LogBatch": {
        "type": "object",
        "required": [
          "logs"
        ],
        "properties": {
          "logs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LogEntry"
            }
          }
        }
      },
      "InsertResult": {
        "type": "object",
        "properties": {
          "accepted": {
            "type": "integer",
            "description": "Number of valid log entries stored"
          },
          "rejected": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "index": {
                  "type": "integer",
                  "description": "Position of the rejected entry in the batch"
                },
                "reason": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "Log": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "level": {
            "type": "string"
          },
          "service": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "attributes": {
            "type": "object",
            "nullable": true,
            "additionalProperties": {
              "type": "string"
            }
          }
        }
      },
      "LogQueryResult": {
        "type": "object",
        "properties": {
          "logs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Log"
            }
          },
          "total": {
            "type": "integer",
            "nullable": true,
            "description": "Total matching rows. Null when cursor pagination is used."
          },
          "next_cursor": {
            "type": "string",
            "nullable": true,
            "description": "Opaque base64 cursor for the next page. Omitted/empty when there are no more pages."
          }
        }
      },
      "Bucket": {
        "type": "object",
        "properties": {
          "start": {
            "type": "string",
            "format": "date-time",
            "description": "Bucket start time"
          },
          "group": {
            "type": "string",
            "nullable": true,
            "description": "Grouping value (service or level) when group_by is provided"
          },
          "count": {
            "type": "integer",
            "description": "Number of logs in the bucket"
          }
        }
      },
      "AggregateResult": {
        "type": "object",
        "properties": {
          "buckets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Bucket"
            }
          }
        }
      },
...
      "Notification": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "service": {
            "type": "string",
            "nullable": true
          },
          "level": {
            "type": "string",
            "nullable": true
          },
          "is_read": {
            "type": "boolean"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
      "Success": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          }
        }
      }
    }
  },
  "paths": {
    "/support/chat": {
      "post": {
        "summary": "Ask the AI support assistant",
        "tags": [
          "Support"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "message"
                ],
                "properties": {
                  "message": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Assistant reply",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "reply": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Message is required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "502": {
            "description": "Support agent unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/notifications": {
      "get": {
        "summary": "List notifications",
        "tags": [
          "Notifications"
        ],
        "responses": {
          "200": {
            "description": "All notifications, newest first",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "notifications": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Notification"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Failed to fetch notifications",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/notifications/read-all": {
      "post": {
        "summary": "Mark all notifications as read",
        "tags": [
          "Notifications"
        ],
        "responses": {
          "200": {
            "description": "All notifications marked as read",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Success"
                }
              }
            }
          },
          "500": {
            "description": "Failed to mark notifications as read",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/notifications/{id}/read": {
      "post": {
        "summary": "Mark a notification as read",
        "tags": [
          "Notifications"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Notification ID",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Notification marked as read",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Success"
                }
              }
            }
          },
          "400": {
            "description": "Invalid notification ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Failed to mark notification as read",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/logs": {
      "post": {
        "summary": "Ingest a batch of log entries",
        "description": "Accepts a batch of log entries. Valid entries are stored; invalid ones are reported per entry.",
        "tags": [
          "Logs"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogBatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "At least one entry was accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InsertResult"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body or no entries were valid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Query logs",
        "description": "Fetch logs filtered by time range, service, level and full-text search, with offset/page or cursor pagination.",
        "tags": [
          "Logs"
        ],
        "parameters": [
          {
            "name": "service",
            "in": "query",
            "description": "Exact service name to filter by",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "level",
            "in": "query",
            "description": "Comma-separated list of levels (debug, info, warn, error)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "since",
            "in": "query",
            "description": "Only logs at or after this timestamp",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "until",
            "in": "query",
            "description": "Only logs at or before this timestamp",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "q",
            "in": "query",
            "description": "Full-text search over messages",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of rows to return (default 100)",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 1000,
              "default": 100
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "1-based page number (offset pagination)",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "description": "Opaque base64 cursor from the previous response (cursor pagination)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Matching logs",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogQueryResult"
                }
              }
            }
          },
          "400": {
            "description": "Invalid query parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/logs/retention/run": {
      "post": {
        "summary": "Run the retention policy",
        "description": "Deletes log entries older than the configured retention period.",
        "tags": [
          "Logs"
        ],
        "responses": {
          "200": {
            "description": "Retention run completed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "deleted": {
                      "type": "integer",
                      "description": "Number of log rows deleted"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/logs/aggregate": {
      "get": {
        "summary": "Aggregate logs into time buckets",
        "description": "Returns log counts bucketed by time, optionally grouped by service or level.",
        "tags": [
          "Logs"
        ],
        "parameters": [
          {
            "name": "since",
            "in": "query",
            "required": true,
            "description": "Bucket window start",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "until",
            "in": "query",
            "required": true,
            "description": "Bucket window end",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "bucket",
            "in": "query",
            "required": true,
            "description": "Bucket interval",
            "schema": {
              "type": "string",
              "enum": [
                "1m",
                "5m",
                "1h",
                "1d"
              ]
            }
          },
          {
            "name": "group_by",
            "in": "query",
            "description": "Split buckets by a field",
            "schema": {
              "type": "string",
              "enum": [
                "service",
                "level"
              ]
            }
          },
          {
            "name": "q",
            "in": "query",
            "description": "Full-text search over messages",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "service",
            "in": "query",
            "description": "Exact service name to filter by",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "level",
            "in": "query",
            "description": "Comma-separated list of levels (debug, info, warn, error)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Aggregated buckets",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AggregateResult"
                }
              }
            }
          },
          "400": {
            "description": "Invalid query parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}